refactored S3 Parameters
