+++ src/s3_parameters.py
@@ -1,0 +1,58 @@
+# Copyright 2024 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Provides S3 Parameters configuration."""
+
+from typing import Any, Optional
+
+from pydantic import BaseModel, Field, validator
+
+
+class S3Parameters(BaseModel):
+    """Configuration for accessing S3 bucket.
+
+    Attributes:
+        access_key: AWS access key.
+        secret_key: AWS secret key.
+        region: The region to connect to the object storage.
+        bucket: The bucket name.
+        endpoint: The endpoint used to connect to the object storage.
+        path: The path inside the bucket to store objects.
+        s3_uri_style: The S3 protocol specific bucket path lookup type. Can be "path" or "host".
+        addressing_style: S3 protocol addressing style, can be "path" or "virtual".
+    """
+
+    access_key: str = Field(alias="access-key")
+    secret_key: str = Field(alias="secret-key")
+    region: Optional[str]
+    bucket: str
+    endpoint: Optional[str]
+    path: str = Field(default="")
+    s3_uri_style: str = Field(alias="s3-uri-style", default="host")
+
+    @validator("endpoint", always=True)
+    @classmethod
+    def check_endpoint_or_region_set(cls, endpoint: str, values: dict[str, Any]) -> str:
+        """Validate that either region or endpoint is set.
+
+        Args:
+            endpoint: endpoint attribute
+            values: all attributes in S3 configuration
+
+        Returns:
+            value of the endpoint attribute
+
+        Raises:
+            ValueError: if the configuration is invalid.
+        """
+        region = values.get("region")
+        if not region and not endpoint:
+            raise ValueError('one of "region" or "endpoint" needs to be set')
+        return endpoint
+
+    @property
+    def addressing_style(self) -> str:
+        """Translates s3_uri_style to AWS addressing_style."""
+        if self.s3_uri_style == "path":
+            return "path"
+        return "virtual"
